requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
attrs==22.2.0
Automat==22.10.0
beautifulsoup4==4.11.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
constantly==15.1.0
contourpy==1.0.7
cryptography==39.0.0
cssselect==1.2.0
cycler==0.11.0
decorator==5.1.1
Django==4.1.5
fake-useragent==1.1.1
filelock==3.9.0
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
google-auth==2.16.0
google-cloud==0.34.0
google-pasta==0.2.0
grpcio==1.51.1
h5py==3.7.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
joblib==1.2.0
keras==2.11.0
kiwisolver==1.4.4
libclang==15.0.6.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.3
numpy==1.24.1
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
panda==0.3.1
pandas==1.5.3
parsel==1.7.0
Pillow==9.4.0
Protego==0.2.1
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==23.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-scrapyd-api==2.1.2
pytz==2022.7.1
pywin32==305
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
Scrapy==2.7.1
scrapyd==1.3.0
seaborn==0.12.2
service-identity==21.1.0
six==1.16.0
sklearn==0.0.post1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
threadpoolctl==3.1.0
tldextract==3.4.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
validators==0.20.0
w3lib==2.1.1
Werkzeug==2.2.2
wrapt==1.14.1
zope.interface==5.5.2
